+++ config/config.lua
@@ -16,6 +16,29 @@ return {
     -- Optional: if you want to attempt su-copy into app storage, set appAutoExecutePath to a writable path (requires root)
     -- appAutoExecutePath = "/data/data/<package>/files/autoexecute",
     instances = {
-
+        [1] = {
+        id = 3,
+        name = "https://www.roblox.com/games/107778070777162",
+        package = "com.apengjers.v5",
+        privateServer = "https://www.roblox.com/games/107778070777162",
+        },
+        [2] = {
+        id = 4,
+        name = "https://www.roblox.com/games/107778070777162",
+        package = "com.apengjers.v6",
+        privateServer = "https://www.roblox.com/games/107778070777162",
+        },
+        [3] = {
+        id = 1,
+        name = "https://www.roblox.com/games/107778070777162",
+        package = "com.apengjers.v3",
+        privateServer = "www.roblox.com/games/107778070777162/Steal-An-Egg",
+        },
+        [4] = {
+        id = 2,
+        name = "https://www.roblox.com/games/107778070777162",
+        package = "com.apengjers.v4",
+        privateServer = "https://www.roblox.com/games/107778070777162",
+        },
     }
 }